Don't blow up if somebody deletes the app directory before an uninstall

+++ src/Shimmer.Core/Utility.cs
@@ -175,6 +175,12 @@ namespace Shimmer.Core
         {
             var di = new DirectoryInfo(directoryPath);
 
+            if (!di.Exists) {
+                LogManager.GetLogger(typeof(Utility))
+                    .Warn("DeleteDirectoryAtNextReboot: does not exist - {0}", directoryPath);
+                return;
+            }
+
             // NB: MoveFileEx blows up if you're a non-admin, so you always need a backup plan
             di.GetFiles().ForEach(x => safeDeleteFileAtNextDir(x.FullName));
             di.GetDirectories().ForEach(x => DeleteDirectoryAtNextReboot(x.FullName));
